Update home1.html
This is a secondary "home" page, the page where any new user would first land before signing up to become a CS client. I may decide to utilize this version or at least a part of it. There are some tweaks that would have to happen before I signed off fully on it though.